Global error handling progress;

+++ MVCApp/Repository/PersonRepository.cs
@@ -13,8 +13,6 @@ namespace MVCApp.Repository
     public class PersonRepository
     {
         private readonly ApplicationDbContext _context;
-        private static readonly log4net.ILog Log = log4net.LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
-
         public PersonRepository()
         {
             _context = new ApplicationDbContext();
@@ -27,79 +25,37 @@ namespace MVCApp.Repository
 
         public Person Get(int id)
         {
-            try
-            {
-                return _context.Persons.FirstOrDefault(x => x.Id == id);
-            }
-            catch (DbException e)
-            {
-                Log.Error($"Error Occured while retrieving a Person with id {id}: {e.Message}");
-                return null;
-            }
-
+            return _context.Persons.FirstOrDefault(x => x.Id == id);
         }
 
         public List<string> GetAllNames()
         {
-            try
-            {
-                return _context.Persons.Select(x => x.Name).ToList();
-            }
-            catch (DbException e)
-            {
-                Log.Error($"Error Occured while retrieving all person names: {e.Message}");
-                return null;
-            }
+            return _context.Persons.Select(x => x.Name).ToList();
         }
 
         public void Add(Person person)
         {
-            try
-            {
-                _context.Persons.Add(person);
-                _context.SaveChanges();
-            }
-            catch (DbException e)
-            {
-                Log.Error($"Error Occured while adding an entry to Person with name {person.Name}: {e.Message}");
-            }
+            _context.Persons.Add(person);
+            _context.SaveChanges();
         }
 
 
         public bool Update(Person person)
         {
             var oldPerson = Get(person.Id);
-            if (oldPerson != null)
-            {
-                try
-                {
-                    _context.Persons.AddOrUpdate(person);
-                    _context.SaveChanges();
-                }
-                catch (DbException e)
-                {
-                    Log.Error($"Error Occured while updating Person with id: {person.Id}: {e.Message}");
-                    return false;
-                }
-                return true;
-            }
-            return false;
+            if (oldPerson == null) return false;
+            _context.Persons.AddOrUpdate(person);
+            _context.SaveChanges();
+            return true;
         }
 
         public bool Delete(int id)
         {
-            try
-            {
-                var person = _context.Persons.FirstOrDefault(x => x.Id == id);
-                _context.Persons.Remove(person);
-                _context.SaveChanges();
-                return true;
-            }
-            catch (DbException e)
-            {
-                Log.Error($"Error Occured while deleting an entry from Person with id {id}: {e.Message}");
-                return false;
-            }
+            var person = _context.Persons.FirstOrDefault(x => x.Id == id);
+            if (person == null) return false;
+            _context.Persons.Remove(person);
+            _context.SaveChanges();
+            return true;
         }
 
         public List<Person> FindAll(Func<Person, bool> predicate)
